Move code that's not charm specific to lib.archive_auth_mirror
